requirements.txt
streamlit==1.12.0
youtube-transcript-api==0.6.2
openai==1.6.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==1.5.3
altair==4.2.2
httpx==0.24.1
